+++ src/users.rs
@@ -6,15 +6,13 @@ use super::errors::{extract_google_api_error, Result};
 
 use super::sessions::{user, service_account};
 use serde::{Deserialize, Serialize};
-use serde_json::json;
 
 use crate::FirebaseAuthBearer;
-use reqwest::Client;
 
 /// A federated services like Facebook, Github etc that the user has used to
 /// authenticated himself and that he associated with this firebase auth account.
 #[allow(non_snake_case)]
-#[derive(Debug, Default, Deserialize, Serialize)]
+#[derive(Debug, Default, Deserialize)]
 pub struct ProviderUserInfo {
     pub providerId: String,
     pub federatedId: String,
@@ -24,7 +22,7 @@ pub struct ProviderUserInfo {
 
 /// Users id, email, display name and a few more information
 #[allow(non_snake_case)]
-#[derive(Debug, Default, Deserialize, Serialize)]
+#[derive(Debug, Default, Deserialize)]
 pub struct FirebaseAuthUser {
     pub localId: Option<String>,
     pub email: Option<String>,
@@ -53,15 +51,9 @@ pub struct FirebaseAuthUserResponse {
 }
 
 #[allow(non_snake_case)]
-#[derive(Debug, Default, Serialize)]
+#[derive(Serialize)]
 struct UserRequest {
     pub idToken: String,
-}
-
-impl UserRequest {
-    fn new(id_token: String) -> UserRequest {
-        UserRequest { idToken: id_token }
-    }
 }
 
 #[inline]
@@ -72,43 +64,22 @@ fn firebase_auth_url(v: &str, v2: &str) -> String {
     )
 }
 
-#[inline]
-fn user_info_internal(
-    auth: String,
-    api_key: &str,
-    firebase_user_id: &str,
-) -> Result<FirebaseAuthUserResponse> {
-    let url = firebase_auth_url("lookup", api_key);
-
-    let mut resp = Client::new()
-        .post(&url)
-        .json(&UserRequest::new(auth))
-        .send()?;
-
-    extract_google_api_error(&mut resp, || firebase_user_id.to_owned())?;
-
-    Ok(resp.json()?)
-}
-
 /// Retrieve information about the firebase auth user associated with the given user session
 ///
 /// Error codes:
 /// - INVALID_ID_TOKEN
 /// - USER_NOT_FOUND
 pub fn user_info(session: &user::Session) -> Result<FirebaseAuthUserResponse> {
-    user_info_internal(session.access_token(), &session.api_key, &session.user_id)
-}
+    let url = firebase_auth_url("lookup", &session.api_key);
 
-#[inline]
-fn user_remove_internal(auth: String, api_key: &str, firebase_user_id: &str) -> Result<()> {
-    let url = firebase_auth_url("delete", api_key);
-    let mut resp = Client::new()
+    let mut resp = session.client()
         .post(&url)
-        .json(&UserRequest::new(auth))
+        .json(&UserRequest { idToken: session.access_token() })
         .send()?;
 
-    extract_google_api_error(&mut resp, || firebase_user_id.to_owned())?;
-    Ok({})
+    extract_google_api_error(&mut resp, || session.user_id.to_owned())?;
+
+    Ok(resp.json()?)
 }
 
 /// Removes the firebase auth user associated with the given user session
@@ -117,19 +88,47 @@ fn user_remove_internal(auth: String, api_key: &str, firebase_user_id: &str) -> 
 /// - INVALID_ID_TOKEN
 /// - USER_NOT_FOUND
 pub fn user_remove(session: &user::Session) -> Result<()> {
-    user_remove_internal(session.access_token(), &session.api_key, &session.user_id)
+    let url = firebase_auth_url("delete", &session.api_key);
+    let mut resp = session.client()
+        .post(&url)
+        .json(&UserRequest { idToken: session.access_token() })
+        .send()?;
+
+    extract_google_api_error(&mut resp, || session.user_id.to_owned())?;
+    Ok({})
 }
 
 
 #[allow(non_snake_case)]
 #[derive(Default, Deserialize)]
-struct CreateUserResponse {
-    pub localId: String,
-    pub idToken: String,
-    // access token
-    pub refreshToken: String,
-    pub email: String,
+struct SignInUpUserResponse {
+    localId: String,
+    idToken: String,
+    refreshToken: String,
 }
+
+#[allow(non_snake_case)]
+#[derive(Serialize)]
+struct SignInUpUserRequest {
+    pub email: String,
+    pub password: String,
+    pub returnSecureToken: bool,
+}
+
+fn sign_up_in(session: &service_account::Session, email: &str, password: &str, action: &str) -> Result<user::Session> {
+    let url = firebase_auth_url(action, &session.credentials.api_key);
+    let mut resp = session.client()
+        .post(&url)
+        .json(&SignInUpUserRequest { email: email.to_owned(), password: password.to_owned(), returnSecureToken: true })
+        .send()?;
+
+    extract_google_api_error(&mut resp, || email.to_owned())?;
+
+    let resp: SignInUpUserResponse = resp.json()?;
+
+    Ok(user::Session::new(&session.credentials, Some(&resp.localId), Some(&resp.idToken), Some(&resp.refreshToken))?)
+}
+
 
 /// Creates the firebase auth user with the given email and password and returns
 /// a user session.
@@ -138,20 +137,16 @@ struct CreateUserResponse {
 /// EMAIL_EXISTS: The email address is already in use by another account.
 /// OPERATION_NOT_ALLOWED: Password sign-in is disabled for this project.
 /// TOO_MANY_ATTEMPTS_TRY_LATER: We have blocked all requests from this device due to unusual activity. Try again later.
-pub fn create_user(session: &service_account::Session, email: &str, password: &str) -> Result<user::Session> {
-    let url = firebase_auth_url("signUp", &session.credentials.api_key);
-    let mut resp = Client::new()
-        .post(&url)
-        .json(&json!({
-            "email": email,
-            "password": password,
-            "returnSecureToken": true,
-        }))
-        .send()?;
+pub fn sign_up(session: &service_account::Session, email: &str, password: &str) -> Result<user::Session> {
+    sign_up_in(session, email, password, "signUp")
+}
 
-    extract_google_api_error(&mut resp, || email.to_owned())?;
-
-    let resp: CreateUserResponse = resp.json()?;
-
-    Ok(user::Session::new(&session.credentials, Some(&resp.localId), Some(&resp.idToken), Some(&resp.refreshToken))?)
+/// Signs in with the given email and password and returns a user session.
+///
+/// Error codes:
+/// EMAIL_NOT_FOUND: There is no user record corresponding to this identifier. The user may have been deleted.
+/// INVALID_PASSWORD: The password is invalid or the user does not have a password.
+/// USER_DISABLED: The user account has been disabled by an administrator.
+pub fn sign_in(session: &service_account::Session, email: &str, password: &str) -> Result<user::Session> {
+    sign_up_in(session, email, password, "signInWithPassword")
 }